Made changes to layout and added User History

+++ twitoff/db_model.py
@@ -13,7 +13,7 @@ class User(DB.Model):
         return '<User %r>' % self.username
 
 class Tweet(DB.Model):
-    id = DB.Column(DB.BigInteger, primary_key=True)
+    id = DB.Column(DB.Integer, primary_key=True)
     tweet = DB.Column(DB.String(280), unique=True, nullable=False)
     embedding = DB.Column(DB.PickleType, nullable=False)
     user_id = DB.Column(DB.BigInteger, DB.ForeignKey('user.id'), nullable=False)
